cleaned up the twilio stuff and removed unused files, just need to add CSS to the Challenge page
