--- app/actions/firebase_actions.js
@@ -1,1 +1,0 @@
-//
