--- app/components/Challenge/common/global.css
@@ -1,7 +1,0 @@
-.displayf {
-    display: flex;
-}
-
-.mt1 {
-    margin-top: 1rem;
-}
