--- app/components/Challenge/layout/App.css
@@ -1,16 +1,0 @@
-.form__wrapper {
-    width: 100vw;
-    height: 100vh;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    text-align: center;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-}
-
-.form__container {
-    flex-flow: column wrap;
-}
