--- app/components/WeatherForm.jsx
@@ -1,26 +1,0 @@
-var React = require('react');
-
-var WeatherForm = React.createClass({
-  onFormSubmit: function (e){
-    e.preventDefault();
-
-    var location = this.refs.location.value;
-    if (location.length > 0){
-      this.refs.location.value = '';
-      this.props.onSearch(location);
-    }
-  },
-
-  render: function () {
-    return(
-      <div className="font">
-        <form onSubmit={this.onFormSubmit}>
-          <input type="search" ref="location" placeholder="Search weather by city"/>
-          <button className="button expanded hollow button-form">Get Weather</button>
-        </form>
-      </div>
-    );
-  }
-})
-
-module.exports = WeatherForm;
