--- index.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App2 from './App2';
-import './index.css';
-
-ReactDOM.render(
-  <App2 />,
-  document.getElementById('root')
-);
