+++ webpack.config.js
@@ -75,9 +75,6 @@ module.exports = {
       Examples: 'app/components/Examples.jsx',
 			Login: 'app/components/Login.jsx',
 			Gameroom: 'app/components/Gameroom.jsx',
-      WeatherForm: 'app/components/WeatherForm.jsx',
-      WeatherMessage: 'app/components/WeatherMessage.jsx',
-      openWeatherMap: 'app/api/openWeatherMap.jsx',
       ErrorModal: 'app/components/ErrorModal.jsx',
       applicationStyles: 'app/styles/app.scss',
 			actions: 'app/actions/actions.jsx',
